Added getRate and validate data to the seperate File
