+++ src/page/franchise-order/ShipmentInformation.tsx
@@ -1,17 +1,16 @@
 import { Controller, useFieldArray, useForm } from "react-hook-form";
-
 import { zodResolver } from "@hookform/resolvers/zod";
 import Input from "./components/ui/Input";
 import { BoxMeasurement } from "./components/BoxMeasurement";
 import { Trash2 } from "lucide-react";
 import { Igst } from "./components/Igst";
 import { useEffect, useState } from "react";
-import apiClient from "./api/apiClient";
 import Select from "./components/ui/Select";
 import { ShipmentInformationData } from "./interface";
 import { currencyOptions } from "./config/currencyOptions";
 import { DateComponent } from "@/components/Date";
 import { orderDetailsSchema } from "@/zod/franchiseOrderSchema";
+import { getRate, validateData } from "./utils/shipmentUtils";
 interface IOrderDetailsProps {
   data: ShipmentInformationData;
   onNext: (formData: ShipmentInformationData) => void;
@@ -26,6 +25,7 @@ export const ShipmentInformation = ({ data, onNext }: IOrderDetailsProps) => {
     const pincode = localStorage.getItem("pincode");
     return pincode;
   };
+
   useEffect(() => {
     const updatedPinCode = () => {
       const pincode = getPincode();
@@ -42,6 +42,7 @@ export const ShipmentInformation = ({ data, onNext }: IOrderDetailsProps) => {
     const country = localStorage.getItem("country");
     return country;
   };
+
   useEffect(() => {
     const updatedCountry = () => {
       const country = getCountry();
@@ -64,75 +65,23 @@ export const ShipmentInformation = ({ data, onNext }: IOrderDetailsProps) => {
     resolver: zodResolver(orderDetailsSchema),
     defaultValues: { ...data, invoiceCurrency: "INR" },
   });
+
   const { fields, append, remove } = useFieldArray({
     control,
     name: "items",
   });
-  const currency = watch("invoiceCurrency");
 
+  const currency = watch("invoiceCurrency");
   const total = fields.reduce((acc, _, index) => {
     const qty = watch(`items.${index}.qty`);
     const unitPrice = watch(`items.${index}.unitPrice`);
     return acc + Number(qty) * Number(unitPrice);
   }, 0);
 
-  const getRate = async () => {
-    const currentActualWeight = watch("actualWeight");
-    const currentLength = watch("length");
-    const currentBreadth = watch("breadth");
-    const currentHeight = watch("height");
-    const ratePayload = {
-      customer_shipping_postcode: shippingPincode,
-      customer_shipping_country_code: country,
-      package_weight: currentActualWeight,
-      package_length: currentLength,
-      package_breadth: currentBreadth,
-      package_height: currentHeight,
-    };
-    try {
-      const response = await apiClient.post("/orders/get-shipper-rates", ratePayload);
-      console.log(response.data, "Shipper Rates Response");
-      localStorage.setItem("shipperRates", JSON.stringify(response.data.data));
-      window.dispatchEvent(new Event("storage"));
-      console.log(response.data, "rate from response");
-    } catch (error) {
-      console.error("Error in getting Shipper Rate:", error);
-    }
-  };
-
-  const validateData = async (data: ShipmentInformationData) => {
-    const validatePayload = {
-      csbv: "0",
-      currency_code: data.invoiceCurrency,
-      package_weight: data.actualWeight,
-      package_height: data.height,
-      package_length: data.length,
-      package_breadth: data.breadth,
-      vendor_order_item: data.items.map((item) => ({
-        vendor_order_item_name: item.productName,
-        vendor_order_item_sku: item.sku,
-        vendor_order_item_quantity: item.qty,
-        vendor_order_item_unit_price: item.unitPrice,
-        vendor_order_item_hsn: item.hsn,
-        vendor_order_item_tax_rate: item.igst,
-      })),
-    };
-    try {
-      const response = await apiClient.post("/orders/validate-order-invoice", validatePayload);
-      setResError(null);
-      console.log(response.data, "Validation successful.");
-      return true;
-    } catch (error) {
-      setResError(String(error.response?.data?.message || "Validation error."));
-      console.error("Error validating order invoice:", error);
-      return false;
-    }
-  };
-
   const onSubmit = async (ShipmentInformationData: ShipmentInformationData) => {
-    const isValid = await validateData(ShipmentInformationData);
+    const isValid = await validateData(ShipmentInformationData, setResError);
     if (!isValid) return;
-    await getRate();
+    await getRate(shippingPincode, country, watch("actualWeight"), watch("length"), watch("breadth"), watch("height"));
     onNext(ShipmentInformationData);
   };
   console.log(resError, "resError");
