+++ src/page/franchise-order/config/stepsInfo.ts
@@ -1,0 +1,1 @@
+export const stepTiles = ["consignorDetail", "consigneeDetail", "shipmentInformation", "selectShippingPartner"];
